refactor: move defTagRegex to utils for content script reuse
parser.ts imports settings.ts, which imports the joplin api, so a content
script cannot import from parser.ts without pulling the whole plugin in.
utils.ts has no imports, which is why tagMarkdownItPlugin.ts keeps its own
divergent copy of the pattern rather than importing the canonical one.

Move the declaration to utils.ts so the upcoming editor content script can
share it. The regex literal is byte-identical and settings.ts and the test
are its only consumers, so behaviour is unchanged.

tagMarkdownItPlugin.ts is deliberately left alone: its copy uses the
capture-group form (^|\s)#(...) rather than the lookbehind, and unifying
them would change preview rendering.
